Increase CF_DIAL_TIMEOUT to match CATs
We were seeing i/o timeouts in CATs and upped the timeout to resolve them.  We saw the same errors [in smoke tests recently](https://runtime.ci.cf-app.com/teams/main/pipelines/cf-release/jobs/vsphere-smoke-tests/builds/13) so the solution seems to be to increase the timeout.  I'm upping it in the `bin/test` script here since we don't have a concourse task equivalent for smoke tests and the i/o errors we saw were coming from the bosh errand anyway (so would not be affected by changing the concourse task).