Fix ginkgo compilation error
Ginkgo 1.9 adds a new dependency to `hpcloud/tail`. This dependency was not copied which led to a compilation error for ginkgo.
This only happened when the ginkgo bin was not on `$PATH` and had to be compiled manually.

This commit fixes the issue by copying _all_ dependencies to the current directory.

Signed-off-by: Tobias Zipfel <tobias.zipfel@de.ibm.com>