Stop using dependabot for updating go modules (#80)
The idea is to manage the contents of this file with one tool. We want to automatically bump the version of the golang package in cf-smoke-tests-release, which must track the version of go in this repo's go.mod file (currently updated manually). If the cf-smoke-tests-release pipeline bumps the version of go, it makes sense for it to also bump the other go modules for cf-smoke-tests.