refactor(bin/test.ps1): Remove `CF_DIAL_TIMEOUT` export
This export was added ~6 years ago because it solved some issues in
CATs, but it's no longer used in CATs.

Besides, this script is used directly by the `smoke_test_windows` job in
`cf-smoke-tests-release`, which allows for this env var to be set to a
BOSH property. We shouldn't override that property here.