fix: Dependabot requires bundler v2
Adds a bundler version to assets/curler in order have dependabot use
bundler v2 and avoid the v1 deprecation:
https://github.blog/changelog/2024-09-05-deprecation-notice-dependabot-dropping-support-for-bundler-v1/.

I'm not totally certain that this change will fix dependabot. It's also
going to create a slight ongoing maintenance burden as we'll need to
bump this bundler version with `bundle update --bundler` on occasion.