chore: Replace build_app.sh with Makefile
Replaces build_app.sh with Makefile to make the script more lightweight
and clear.

It seems like the goal of the build_app.sh script is to build a new app
using the GOOS, GOARCH, and Go version associated with the cf-d-c-t
image. While that is somewhat useful its limited because that image
doesn't necessarily reflect the cflinuxfs that the app will be pushed
to. I think that its goals can all be accomplished by specifying the
values as env vars in a Makefile and using the new forward/backward
compatibility in Go. This Makefile also does not depend upon docker and
runs much faster.

Also turned off cgo with `CGO_ENABLED=0` to make the binary static.